SearchForm: move to Search/Form
