--- src/components/search/SearchForm/index.js
@@ -1,2 +1,0 @@
-import SearchForm from './SearchForm';
-export default SearchForm;
